be/requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Redis
redis[hiredis]==5.0.1

# Pydantic
pydantic==2.6.1
pydantic-settings==2.1.0

# WebSockets
websockets==12.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# LangGraph (orchestration)
langgraph==0.0.26

# DSPy (structured LLM)
dspy-ai==2.4.0

# Google Gemini LLM
google-generativeai>=0.3.0

# Async utilities
anyio==4.2.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# Development
black==24.1.1
ruff==0.2.1
